refactor(content): consolidate dialog host UI stack
