+++ tests/unit/content/SupportPrompt.test.ts
@@ -284,6 +284,25 @@ describe('SupportPrompt', () => {
     expect(toastShadow.querySelector('[data-role="github-link"]')).toBeTruthy();
   });
 
+  it('dismisses support toasts from outside pointerdown', async () => {
+    const { SupportPrompt } = await import('../../../src/content/ui/supportPrompt');
+    const prompt = new SupportPrompt(document);
+    await prompt.show({ status: 'success' });
+
+    getPromptHost().shadowRoot?.querySelector<HTMLButtonElement>('[data-role="like-btn"]')?.click();
+    await flushMicrotasks();
+
+    const toastShadow = getToastShadow();
+    const toast = toastShadow.querySelector<HTMLElement>('#aiob-support-toast');
+    expect(toast).toBeTruthy();
+
+    document.body.dispatchEvent(new Event('pointerdown', { bubbles: true }));
+    expect(toast?.classList.contains('is-visible')).toBe(false);
+
+    toast?.dispatchEvent(new Event('transitionend'));
+    expect(document.getElementById('aiob-support-toast-host')).toBeNull();
+  });
+
   it('closes from the Stitch overlay action', async () => {
     const { SupportPrompt } = await import('../../../src/content/ui/supportPrompt');
     const prompt = new SupportPrompt(document);
